+++ PSNovusTools/Private/Invoke-SSHSessionItem.ps1
@@ -1,0 +1,16 @@
+function Invoke-SSHSessionItem([SSH.SshSession]$session, [string]$sshCommand) {
+
+    $sshResult = $null
+
+    try {
+       
+        # Command to start the service and get its status
+        $sshresult = Invoke-SSHCommand -SessionId $session.SessionId -Command $sshCommand
+    }
+    catch {
+        Write-Error "Failed SSH command: $sshCommand Error: $_"
+        $sshresult = $null   
+    }
+
+    return $sshresult
+}